Using ducks which extend `org.paninij.lang.String`.
Minor refactoring of `org.paninij.lang.String` itself to make it
compatible with `ResolvableFuture` and `Panini$Message`.

The generation of ducks of the form `org_paninij_lang_String$Duck$*`
is currently handled as a special case. This special case is flagged
using `Category.PANINICUSTOM`. `examples.duckshapes.PaniniCustom` was
created for testing.