Refactored `MakeCapsuleInterface` to use helpers.
Implementation of `MakeCapsuleInterface` was simplified by using helper
methods defined in `Source` and the new `PaniniModelInfo` class. Also
renamed `ModelInfo` -> `JavaModelInfo`.

These two `ModelInfo` classes hold util/helper methods for finding out
properties of elements (and maybe in the future other objects). Any
properties which are universal to Java go in `JavaModelInfo`, whereas
properties which are specific to the Panini programming model go in
`PaniniModelInfo`. For example, the predicate `needsProcedureWrapper()`
goes in `PaniniModelInfo`.