Removed `MakeSignature`; using user's originals.
Previously, we were generating a `*$Signature` artifact from every
user-defined class annotated with `@Signature`. It seems that this is no
longer necessary. The user-defined signatures can be used directly by
the user in his/her template classes and also in any auto-generated
artifacts (e.g. capsules which implement a signature).