Design args no longer used.
Previously, the arguments of a capsule template's design declaration
were used to infer which of the capsule's fields were either

- wired capsules
- child capsules
- state variables

We discovered that our initial attempts to make this inferrence did not
work. (See GitHub Issue #31.) To fix this, we added `@Child`- and
`@Wired`-annotations. This made the list of capsules arguments to a
template's `design()` declaration unnecessary and the process of
delegating to that method more complicated than it needed to be.

All of these complications were removed, and the helper methods for
identifying which capsule fields are `@Wired` and `@Child` are much
simpler.

This code has been tested with the `HelloWorldTemplate` example.

Note that the `self` argument is still *required*.