Revised `formatAligned()` when zero lines given.
Previously, `Source.formatAligned()` threw an exception when there were
no lines given to it. However, this makes certain cases where the
function is used unnecessarily complicated. The revised behavior--to
simply remove the "##" from the format string--is better because it
means that the user does not need to check whether a list of `lines` is
empty before passing that list into `Source.formatAligned()`.