Adds ownership transfer check at future resolution.
This adds a first-draft of a naive runtime ownership check at the point
where a server capsule resolves a future for use by a client. This check
is meant to guarantee that only objects which are un-owned by the server
capsule are being transferred to the client capsule.

This adds two classes in `org.paninij.examples.ownership` to serve as a
basic test for these checks.