Bugfix: `panini$getAllState() gave wrong obj graph.
Previously, a capsule's `panini$getAllState()` method was giving out the
entire `panini$encapsulated` object. This was causing an extremely
critical a performance bug. In addition to a capsule's state variables,
`panini$encapsulated` holds references a capsule's @Child and @Wired
references. This means that the explorer was traversing more than just
a capsule's local state. It was also traversing the state of other
capsules!

Interestingly, the results produced by the check were still correct,
since an outgoing transfer must be disjoint with all of the object
graphs of all the non-local capsules as well!.