Drafted capsule tester generation.
These are special capsules that can be used to make testing the behavior
of capsule simpler (both for capsule-generation developers and capsule
developers). The syntax is inspired by JUnit tests. Each test defined in
a capsule tester is run independently of the others. For an example, see
`org.paninij.apt.test.BasicCapsuleTester`.

In the current implementation, each class annotated with @CapsuleTester
is used as a capsule template for generating a passive thread-capsule.
These generated capsules are *almost* the same as if they had been
generated in the normal way, with just a few small exceptions, most
importantly,

- Naming conventions have been modified.
- Each tester capsule has been given a `main()` method.

The `main()` method is the key here. The main method essentially
contains a `for` loop that instantiates a new tester capsule instance
for each of the tests. On each of these instances a single test
invocation message is sent and a single shutdown message is sent.

There are currently many deficiencies to this proof of concept, but I am
hoping that it is a good enough hack that we can hold off on cleaning it
up until we make capsule generation model-based.