Capsule errors stored in `Panini$System.errors`.
To implement `BasicOwnershipTester.assertThrowsAssertionError()`, one
capsule needed to be made aware that an assertion error was thrown
within another capsule. `Panini$System.errors` was implemented to enable
this.

The `run()` method of a "normal" capsule, will catch all throwables
which that capsule throws during its execution. Another capsule, such as
`BasicOwnershipTester` is able to see whether exceptions have been
thrown.

This solution is far from ideal. It only meant to be a workaround for
use in testing capsules.