Drafted experiment/test for WALA class loading.
This was done in order to try to track down a bug where the
`LeakyServer` capsule interface was not being loaded. However, that bug
seems to be gone now. It might have had something to do with maven or
the build process. Maybe it was just that while I was running the
Eclipse debugger there really wasn't a compiled version of the iterface
within `target/test-classes`. I have no idea how that could be, unless I
just hadn't run `maven clean install` for a *really* long time.