Adds entrypoint class which auto-adds dummies.
It was previously necessary to instantiate the `@Child` and `@Wired`
fields of a capsule template manually, but with this commit, the CG
construction now works without this workaround. This is achieved using
the new `CapsuleTemplateEntrypoint` class, which, for a given entrypoint
method, instantiates a capsule template receiver instance whose capsule
fields automatically contain capsule dummy instances. An entrypoint does
this by adding instructions to the fake root method.