Deleted CG tests; `apt` can now make CG PDFs.
A `PaniniProcessor` can now be made to generate call graph PDFs using
the `-Apanini.soter.callGraphPDFs` annotation processor option. The
option's value will be the output directory for the PDFs. In particular,
if the option is set, then for each template generated by the processor,
a call graph PDF will be put in this directory.