BUGFIX: Checking transfer safety more consistently.
Previously there were slightly different criteria for checking whether
a transfer's type was known to be safe, depending upon whether the
transfer site was an INVOKE or a RETURN. Now, the helper function
`isKnownSafeTypeForTransfer()` is used more consistently.

Note that that function still needs to be fleshed out with additional
known safe types.