Adds `ArtifactFiler`; it implements `ArtifactMaker`.
This new `ArtifactFiler` complements the `ArtifactCompiler`. The former
simply takes artifacts and files them. The latter files them and also
compiles them. `UserArtifact` was also introduced for use with the
user-defined capsule templates which need to be compiled (but not filed)
by the `ArtifactCompiler` and completely ignored by the `ArtifactFiler`.