BUGFIX: Many non-capsule `@Wired` are instantiated.
Previously, there was a bug in the SOTER analysis where all `@Wired`
decls were assumed to be template interfaces. This used to be true, but
has changed in recent versions of `@PaniniJ`. Now `@Wired` decls of
non-capsule types are instantiated in the fake root node or properly
ignored.

Note that support for `@Wired` arrays has still not yet been added.