Now (no instrumentation) soter test runs via JUnit.
Previously, the basic no instrumentation test was integrated with the
build process of the `at-paninij-soter-tests` Maven project. Now, the
executable is instead invoked within a JUnit test. This is simpler,
allows for assertions within the soter codebase to be to be enabled, and
(most importantly) makes it possible to run this test via the debugger.