Add `at-paninij-soter-example` Maven subproject.
This is mean to serve as an example of how to integrate the `soter` tool
into a Maven build of @PaniniJ capsule templates. The `soter` main
executable is triggered as part of the Maven build. This functionality
was previously integrated with `at-paninij-soter-tests`, but it has been
moved here to simplify and improve the process of testing `soter` (i.e.
Eclipse debugging support).