Move some reusable JSON creation code to own class.
Previously, only the `SoterAnalysis` was creating JSON objects for
logging and debugging purposes. JSON creation was done via a private
(non-static) inner class of `SoterAnalysis`. In order to make be able to
share some of this JSON creation code, some of the code was moved out
into a new abstract class: `AnalysisJsonResultsCreator`.

This new class is meant to be extended for use by various analyses as
they construct JSON (probably with inner classes, as is done with
`SoterAnalysis` currently).