BUGFIX: Template receivers now live in root sites.
Previously, there were many cases where there a `PointerKey` to a root
capsule template was not being marked as live. Now, the transfer site
live analysis marks the capsule template receiver as live for every
"relevant" analysis site defined in the root.

This guarantees that during the later (interprocedural) call graph live
analysis, the fact that the capsule template is live will be propagated
forward correctly.