Move `MethodInstrumenter` to own compilation unit.
This (non-static) inner class was moved to its own class using an
Eclipse auto-refactoring. This was done as a first step to refactor this
class to have better separation of concerns so that it can be re-used by
another (as yet unimplemented) class which instruments all transfer
sites, regardless of the results of the `SoterAnalysis`.