Change instrumenter factory for better reuse.
Previously, the `SoterInstrumenterFactory` was just used to construct
`SoterInstrumenter` instances. However, the only substantive thing which
this factory did was create a `ClassInstrumenter` for use with a
`SoterInstrumenter`.

In order to make these `ClassInstrumenter` instances usable by other
`Instrumenter` types, `SoterInstrumenterFactory` has been changed into
`TemplateInstrumenterFactory` which makes these `ClassInstrumenter`
instances.

This revised factory type will be able to be used in an upcoming
instrumenter which instruments all transfer sites.