Make `private` fields illegal.
This was done so that a capsule is able to access fields of its
encapsulated template instance. This prevents certain strange user-
facing errors in the case that a capsule template declares a private
field of reference type (i.e. an object or an array).

These errors would happen because the code generated for
`panini$getAllState()` directly references those fields of a capsule
template of reference type (i.e. an object or array).

I believe that this case of `panini$getAllState()` is currently the only
one in which the fields of a capsule template are referenced from
outside of the capsule itself (i.e. in generated code).