Remove `at-paninij-runtime` as dep.
This was motivated by errors seen when trying to compile the project
within IntelliJ IDEA: The definitions in `runtime` were not being found
and this was solved by simply removing the `<scope>` elements.

However, these explicit runtime dependencies are now redundant, since
`runtime` is included in `proc` via the `maven-shade-plugin`.