Add simple `JCTree` access proof-of-concept
This commit is a proof of concept to demonstrate that `JCTree` instances
can be accessed from within the @PaniniJ annotation processor. All
root compilation units are looked-up in the round environment, casted
to `JCTree.JCCompilationUnit`, and printed out.

This change required the addition of the `com.sun.tools` as a *system*
dependency of `at-paninij-proc`. This is a bit more brittle than I'd
like, but it seem like the right choice for now.