Replace `:core:proc-tests` with a new system
Details on the design and usage of this new system are given in the
javadoc of `org.paninij.proc.ProcTests`.

The main advantage of this system over the last one is that it lets the
test writer write a small file--one suffixed by `compile.properties--to
describe a compilation task to be performed. Unlike `:core:proc-test`,
it is now much easier to create compilation tasks which run in relative
isolation from one another.

This implementation takes much inspiration from the recent testing
overhaul by @jlmaddox. However, this is a re-implementation of that
system which I believe improves on Jackson's work in a few ways:

- The `compile.properties` file format is better specified and more
closely matches a Java compiler task's parameters.
- There is now a 1-1 correspondence between a `compile.properties` and
a test compilation task run by the system.
- Individual testing compilation tasks now derive their names from their
`compile.properties` file in an intuitive way, and these names are
displayed when running the tests (an improvement over the previous
implementation which just indexed all of the cases with a number).

Additionally, this implementation makes for a slightly simpler @PaniniJ
codebase, because:

- It now uses the library `me.dwtj:java-compiler-utils` (which I've
published to Maven Central) to construct compilation tasks.
- It eliminates the `:core:proc-tests` subproject.