Deploy paninij site from Travis to GitHub Pages
Deployment itself is performed by pushing to the @PaniniJ organization's
GitHub pages repo, that is, `paninij/paninij.github.io`. Authentication
of this push is done via this repository's current deployment key-pair,
whose fingerprint is

    SHA256:XyaXduOGMrF6kEEtVyBVHzCcPrZNGPpiXdUBlIF/nEw GitHub Pages Deploy Key for Travis

whose full public key has been added as a file in this repo, and whose
private key is an encrypted file in the repo. The private key is
encrypted using Travis's encrypt-file feature. (For info on this
feature, see `https://docs.travis-ci.com/user/encrypting-files/`.)

Currently, only the javadocs of `org.paninij.lang` are published to this
site. Future commits will expand the content on this site.