Split processor to fix #143 #152; rename checks
This is a major overhaul of the annotation processor. What was
the `PaniniProcessor` has now been split into two processors:

- org.paninij.proc.RoundZeroProcessor
- org.paninij.proc.RoundOneProcessor

See the Javadoc of the `org.paninij.proc` package for background and
motivation.

Note that three of the processor's compile tests still fail, since checks
for these have not yet been implemented:

- org.paninij.proc.check.capsule.duckability.NoReturnTypeTemplate
- org.paninij.proc.check.capsule.fields.recursive.BadTemplate
- org.paninij.proc.check.capsule.fields.recursive.CycleBad