Disabled `:benchmarks:jmh-harness`; see issue #157
This subproject isn't being used, and it currently doesn't work because
of some obscure error. On my system, the `:benchmarks:jmh-harness:jmhJar`
task seems to always fail with the message

    Could not expand ZIP '/Users/dwtj/.gradle/caches/modules-2/files-2.1/org.scala-lang.modules/scala-xml_2.11/1.0.2/820fbca7e524b530fdadc594c39d49a21ea0337e/scala-xml_2.11-1.0.2.jar'.

The relevant portion of the stacktrace seems to be

    org.gradle.api.file.DuplicateFileCopyingException: Encountered duplicate path "scala/xml/Atom.class" during copy operation configured with DuplicatesStrategy.FAIL

though I don't know how to address this problem at the moment.