Add `:site:htmlProofer` task.
This task is run as part of `:site:check` to look for a number of basic
problems with the generated static site.

This task is currently configured to ignore problems in the Javadoc.
This was done because there are just too many problems with the javadoc
HTML generated for `org.paninij.lang.String`. Mostly, these appear to be
broken links to classes and methods that are in the standard library.

Also added some descriptions for the custom tasks in `:site`.