Draft preliminary runtime ownership move check
The current naive implementation works by traversing the entire object
graph reachable from the moved object and then traversing the entire
object graph reachable from the sender. If an object is found to be in
both, then an error is thrown.

See the agent's `it2` for a trivial example which makes this error
happen.