Refactor agent tests to use Gradle, not CTest
This change is useful because (1) the implementation is simpler, (2) the
@PaniniJ annotation processor is now run when an integration test is
compiled, and (3) it is now much easier to attach to tests with a
debugger.

IDE-debuggable integration test can now be launched from within my IDE
or from the command line (using Gradle's `--debug-jvm` switch). Using
either of these methods, one can use the "double debugger" method often
used in JNI development, where both a Java debugger and a native
debugger is attached to a `java` process. In this way, one can step
through both Java code and native code running in a process.