Draft agent's `check_reachable_from_stack()`
Local variables which reach a moved variable can even be detected on in
stack frames older than the stack frame in which the move is being
performed.  (See `it3` for an example.)

This initial proof-of-concept does not yet distinguish between live and
dead variables.  This makes it entirely unusable. (See `it1` for an
example of its current overzealousness.)