Tweak which modifiers are considered (il)legal
See this commit's modifications to

    org.paninij.proc.check.capsule.CheckForIllegalModifiers

for details about which modifiers became legal/illegal. All example
capsule cores in the codebase were modified accordingly. There appear to
be no build errors associated with this change.

This commit was inspired by Issue #153, but it does not quite implement
it. This commit makes it optional to have a public modifier on a
procedure. Unfortunately, we can't just make it illegal to use public,
since we need public in the case that a procedure is part of a
signature implementation. Capsule cores in the codebase were modified to
use the preferred option, that is, they were changed to not use public
on their procedures (unless necessary).